career-compass-multi-agent/career-compass-agents/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
strands-agents>=1.0.0
strands-agents-tools>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
